+++ src/app/communities/social.model.ts
@@ -1,0 +1,4 @@
+export class Social {
+  public name: string;
+  public url: string;
+}
