--- src/app/communities/social.ts
@@ -1,4 +1,0 @@
-export class Social {
-  name: string;
-  url: string;
-}
